% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{phist_cache}
\alias{phist_cache}
\title{plothistory Package Cache}
\usage{
phist_cache()
}
\value{
The \code{phist_cache()} function returns an instance of class
\code{BiocFileCache}.
}
\description{
Function to access (and manage) the plothistory package
cache. \code{phist_chache()} returns the central \code{plothistory} cache
directory.

When plots are cached, they are named after their md5sum.
}
\examples{

phist_cache()
}
\author{
Laurent Gatto
}
