% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phdir.R
\name{phist_dir}
\alias{phist_dir}
\alias{phist_tmp_dir}
\alias{phist_current_dir}
\alias{phist_cache_dir}
\title{Create a Plot History Directory}
\usage{
phist_dir(path, ...)

phist_tmp_dir()

phist_current_dir()

phist_cache_dir()
}
\arguments{
\item{path}{\code{character(1)} with a single path name, passed to
\code{\link[=dir.create]{dir.create()}}.}

\item{...}{Additional parameters passed to \code{\link[=dir.create]{dir.create()}}.}
}
\value{
Used for its side effect of creating a
directory. Invisibly returns the path to the newly created
directory.
}
\description{
The directory creating is delegated to \code{\link[=tempdir]{tempdir()}} by
\code{phist_tmp_dir()} and \code{\link[=dir.create]{dir.create()}} by \code{phist_dir()}.
}
