% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothistory.R
\name{plothistory}
\alias{plothistory}
\title{Record Your Plot History}
\usage{
plothistory(phdir = phist_tmp_dir(), ...)
}
\arguments{
\item{phdir}{\code{character(1)} defining the plothistory
directory. Default is to \code{\link[=phist_tmp_dir]{phist_tmp_dir()}}.}

\item{...}{additional parameters passed to \code{\link[httpgd:hgd]{httpgd::hgd()}}.}
}
\value{
The value of the error code (0 for success) of the
background script.
}
\description{
The \code{plothistory()} function initiates saving all
plots generated during the session in \code{phdir}. It works as
follows:
\itemize{
\item It first starts a web server graphics device \code{\link[httpgd:hgd]{httpgd::hgd()}}.
\item Next it creates and runs a websocket connected to the graphics
device.
\item Every time the user creates a plot, the web socket receives an
event that triggers the saving of the plot into the user defined
directory \code{phdir}.
}
}
