% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothistory.R
\name{plothistory}
\alias{plothistory}
\title{Record Your Plot History}
\usage{
plothistory(phdir = phist_tmp_dir())
}
\arguments{
\item{phdir}{\code{character(1)} defining the plothistory
directory. Default is to \code{\link[=phist_tmp_dir]{phist_tmp_dir()}}.}
}
\value{
Invisibly returns \code{TRUE}
}
\description{
Record Your Plot History
}
